Wrote a query to create the table work_experience_info

+++ src/main/java/org/example/jobsearch/models/WorkExperienceInfo.java
@@ -12,7 +12,7 @@ import lombok.*;
 public class WorkExperienceInfo {
     private Long id;
     private Long resumeId;
-    private Byte years;
+    private Integer years;
     private String companyName;
     private String position;
     private String responsibilities;
